my_model/code/requirements.txt
torch
torchvision
boto3
Pillow
matplotlib